# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CafeHour {
  day: string;
  hours: string;
}

export interface CafeProfile {
  name: string;
  tagline: string;
  description: string;
  address: string;
  mapUrl: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  instagramUrl: string | null;
  hours: CafeHour[];
}

export type MenuItemStatus = typeof MenuItemStatus[keyof typeof MenuItemStatus];


export const MenuItemStatus = {
  draft: 'draft',
  published: 'published',
  unavailable: 'unavailable',
} as const;

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  category: string;
  price: number;
  currency: string;
  featured: boolean;
  status: MenuItemStatus;
  dietaryLabels: string[];
}

export interface CafeProfileInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  tagline: string;
  /** @minLength 2 */
  description: string;
  /** @minLength 2 */
  address: string;
  mapUrl: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  instagramUrl: string | null;
  hours: CafeHour[];
}

export type MenuItemInputStatus = typeof MenuItemInputStatus[keyof typeof MenuItemInputStatus];


export const MenuItemInputStatus = {
  draft: 'draft',
  published: 'published',
  unavailable: 'unavailable',
} as const;

export interface MenuItemInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  description: string;
  /** @minLength 2 */
  category: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  currency: string;
  featured: boolean;
  status: MenuItemInputStatus;
  dietaryLabels: string[];
}

export type MenuItemUpdate = MenuItemInput;

export interface ReservationInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 6 */
  phone: string;
  /** @nullable */
  email?: string | null;
  date: string;
  /** @minLength 3 */
  time: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  partySize: number;
  /** @nullable */
  notes?: string | null;
}

export type Reservation = ReservationInput & {
  id: number;
  status: string;
  createdAt: string;
};

export interface InquiryInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 2 */
  subject: string;
  /** @minLength 5 */
  message: string;
}

export type Inquiry = InquiryInput & {
  id: number;
  status: string;
  createdAt: string;
};

export type ReservationStatusUpdateStatus = typeof ReservationStatusUpdateStatus[keyof typeof ReservationStatusUpdateStatus];


export const ReservationStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  declined: 'declined',
  completed: 'completed',
} as const;

export interface ReservationStatusUpdate {
  status: ReservationStatusUpdateStatus;
}

export type InquiryStatusUpdateStatus = typeof InquiryStatusUpdateStatus[keyof typeof InquiryStatusUpdateStatus];


export const InquiryStatusUpdateStatus = {
  new: 'new',
  read: 'read',
  archived: 'archived',
} as const;

export interface InquiryStatusUpdate {
  status: InquiryStatusUpdateStatus;
}

export interface OwnerSummary {
  pendingReservations: number;
  newInquiries: number;
  publishedMenuItems: number;
  totalMenuItems: number;
}

export interface ErrorResponse {
  error: string;
}