# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: cafe
    description: Public Corner Cafe information
  - name: menu
    description: Public cafe menu
  - name: reservations
    description: Reservation requests
  - name: inquiries
    description: Customer inquiries
  - name: owner
    description: Protected owner dashboard operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /cafe:
    get:
      operationId: getCafeProfile
      tags: [cafe]
      summary: Get the public cafe profile
      responses:
        "200":
          description: Cafe profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CafeProfile"
  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List visible menu items
      responses:
        "200":
          description: Visible menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
  /reservations:
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Submit a reservation request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Reservation request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          description: Invalid reservation request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inquiries:
    post:
      operationId: createInquiry
      tags: [inquiries]
      summary: Send a message to the cafe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryInput"
      responses:
        "201":
          description: Inquiry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
        "400":
          description: Invalid inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /owner/summary:
    get:
      operationId: getOwnerSummary
      tags: [owner]
      summary: Get the owner dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerSummary"
  /owner/cafe:
    patch:
      operationId: updateOwnerCafeProfile
      tags: [owner]
      summary: Update the public cafe profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CafeProfileInput"
      responses:
        "200":
          description: Updated cafe profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CafeProfile"
  /owner/menu:
    get:
      operationId: listOwnerMenuItems
      tags: [owner]
      summary: List all menu items for the owner
      responses:
        "200":
          description: All menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createOwnerMenuItem
      tags: [owner]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /owner/menu/{id}:
    patch:
      operationId: updateOwnerMenuItem
      tags: [owner]
      summary: Update a menu item
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteOwnerMenuItem
      tags: [owner]
      summary: Delete a menu item
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Menu item deleted
  /owner/reservations:
    get:
      operationId: listOwnerReservations
      tags: [owner]
      summary: List reservation requests
      responses:
        "200":
          description: Reservation requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
  /owner/reservations/{id}:
    patch:
      operationId: updateOwnerReservation
      tags: [owner]
      summary: Update a reservation request status
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationStatusUpdate"
      responses:
        "200":
          description: Updated reservation request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /owner/inquiries:
    get:
      operationId: listOwnerInquiries
      tags: [owner]
      summary: List customer inquiries
      responses:
        "200":
          description: Customer inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inquiry"
  /owner/inquiries/{id}:
    patch:
      operationId: updateOwnerInquiry
      tags: [owner]
      summary: Update an inquiry status
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryStatusUpdate"
      responses:
        "200":
          description: Updated inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CafeHour:
      type: object
      properties:
        day:
          type: string
        hours:
          type: string
      required: [day, hours]
    CafeProfile:
      type: object
      properties:
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        address:
          type: string
        mapUrl:
          type: string
          format: uri
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        hours:
          type: array
          items:
            $ref: "#/components/schemas/CafeHour"
      required:
        - name
        - tagline
        - description
        - address
        - mapUrl
        - phone
        - email
        - instagramUrl
        - hours
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
          format: double
        currency:
          type: string
        featured:
          type: boolean
        status:
          type: string
          enum: [draft, published, unavailable]
        dietaryLabels:
          type: array
          items:
            type: string
      required: [id, name, description, category, price, currency, featured, status, dietaryLabels]
    CafeProfileInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        tagline:
          type: string
          minLength: 2
        description:
          type: string
          minLength: 2
        address:
          type: string
          minLength: 2
        mapUrl:
          type: string
          format: uri
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
          format: email
        instagramUrl:
          type: ["string", "null"]
          format: uri
        hours:
          type: array
          items:
            $ref: "#/components/schemas/CafeHour"
      required: [name, tagline, description, address, mapUrl, phone, email, instagramUrl, hours]
    MenuItemInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
          minLength: 2
        category:
          type: string
          minLength: 2
        price:
          type: number
          minimum: 0
        currency:
          type: string
          minLength: 1
        featured:
          type: boolean
        status:
          type: string
          enum: [draft, published, unavailable]
        dietaryLabels:
          type: array
          items:
            type: string
      required: [name, description, category, price, currency, featured, status, dietaryLabels]
    MenuItemUpdate:
      allOf:
        - $ref: "#/components/schemas/MenuItemInput"
    ReservationInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 6
        email:
          type: ["string", "null"]
          format: email
        date:
          type: string
          format: date
        time:
          type: string
          minLength: 3
        partySize:
          type: integer
          minimum: 1
          maximum: 20
        notes:
          type: ["string", "null"]
      required: [name, phone, date, time, partySize]
    Reservation:
      allOf:
        - $ref: "#/components/schemas/ReservationInput"
        - type: object
          properties:
            id:
              type: integer
            status:
              type: string
            createdAt:
              type: string
              format: date-time
          required: [id, status, createdAt]
    InquiryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        subject:
          type: string
          minLength: 2
        message:
          type: string
          minLength: 5
      required: [name, email, subject, message]
    Inquiry:
      allOf:
        - $ref: "#/components/schemas/InquiryInput"
        - type: object
          properties:
            id:
              type: integer
            status:
              type: string
            createdAt:
              type: string
              format: date-time
          required: [id, status, createdAt]
    ReservationStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, declined, completed]
      required: [status]
    InquiryStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, read, archived]
      required: [status]
    OwnerSummary:
      type: object
      properties:
        pendingReservations:
          type: integer
        newInquiries:
          type: integer
        publishedMenuItems:
          type: integer
        totalMenuItems:
          type: integer
      required: [pendingReservations, newInquiries, publishedMenuItems, totalMenuItems]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]