# lib/api-zod/src/generated/types/cafeProfileInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CafeHour } from './cafeHour';

export interface CafeProfileInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  tagline: string;
  /** @minLength 2 */
  description: string;
  /** @minLength 2 */
  address: string;
  mapUrl: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  instagramUrl: string | null;
  hours: CafeHour[];
}