# lib/api-zod/src/generated/types/inquiryStatusUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InquiryStatusUpdateStatus } from './inquiryStatusUpdateStatus';

export interface InquiryStatusUpdate {
  status: InquiryStatusUpdateStatus;
}