# lib/api-zod/src/generated/types/menuItemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItemInputStatus } from './menuItemInputStatus';

export interface MenuItemInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  description: string;
  /** @minLength 2 */
  category: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  currency: string;
  featured: boolean;
  status: MenuItemInputStatus;
  dietaryLabels: string[];
}