# lib/api-zod/src/generated/types/menuItemUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItemInput } from './menuItemInput';

export type MenuItemUpdate = MenuItemInput;