# lib/api-zod/src/generated/types/ownerSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OwnerSummary {
  pendingReservations: number;
  newInquiries: number;
  publishedMenuItems: number;
  totalMenuItems: number;
}