# lib/api-zod/src/generated/types/reservationStatusUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReservationStatusUpdateStatus } from './reservationStatusUpdateStatus';

export interface ReservationStatusUpdate {
  status: ReservationStatusUpdateStatus;
}